Server/requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
annotated-types==0.7.0
anyio==4.4.0
archspec @ file:///croot/archspec_1709217642129/work
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autolab-core==1.1.1
autolab-perception==1.0.0
backcall==0.2.0
blinker==1.8.2
boltons @ file:///croot/boltons_1677628692245/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cmake==3.25.0
colorlog==6.8.2
comm==0.2.2
conda @ file:///croot/conda_1715635703388/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
ConfigArgParse==1.7
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1714660666131/work
cvxopt==1.3.2
cycler==0.12.1
Cython==3.0.10
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro @ file:///croot/distro_1714488253808/work
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.3
fonttools==4.53.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gradio==3.50.2
gradio_client==0.6.1
grasp_nms==1.0.2
graspnetAPI==1.2.11
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git@df5b48a3efbaa64288d8d0ad09b748ac86f22671
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.16.4
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipython==8.12.3
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lang-sam @ git+https://github.com/luca-medeiros/lang-segment-anything.git@831cdc10906a23aabe4591a4be06f4c989e4ee30
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
lightning==2.3.3
lightning-utilities==0.11.3.post0
lit==15.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1718132535249/work
MinkowskiEngine==0.5.4
mmcv-full==1.3.9
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
open3d==0.18.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==5.22.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648042571233/work
pointnet2==0.0.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
PyWavefront==1.3.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests @ file:///croot/requests_1716902831423/work
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.4.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
supervision==0.18.0
sympy==1.12
tenacity==8.4.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.7
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu118
tqdm @ file:///croot/tqdm_1716395931952/work
traitlets==5.14.3
transformers==4.35.2
transforms3d==0.3.1
trimesh==4.4.1
triton==2.3.1
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3 @ file:///croot/urllib3_1700840488771/work
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.11
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
zstandard @ file:///croot/zstandard_1714677652653/work
